requirements.txt
asgiref==3.5.2
azure-core==1.25.0
azure-storage-blob==12.13.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
cycler==0.11.0
Django==4.1
fonttools==4.34.4
gunicorn==20.1.0
idna==3.3
isodate==0.6.1
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.2
mpld3==0.5.8
msrest==0.7.1
numpy==1.23.1
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pycparser==2.21
pyodbc==4.0.34
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.3.0
tzdata==2022.1
urllib3==1.26.11
whitenoise==6.2.0
plotly==5.6.0
